Cherry-pick for SWDEV-238474
This cherry-pick fixes a memory access fault issue in MIGraphX.

AMDGPU: Fix inlining logic for denormals

This was backwards from intended and missing a test. We perhaps should
just ignored the FP mode here, since it shouldn't be legal to mix code
with different default modes in the absence of strictfp.

Change-Id: I47e284bfa0cc872023eaf09cd7e5726db9b3c39d