SWDEV-321398: replace hostcall module flag with function attribute
This internal version is currently a squash of four upstream reviews:

1. D119087: [AMDGPU] [NFC] refactor the AMDGPU attributor
2. D119308: [AMDGPU] [NFC] Fix incorrect use of bitwise operator.
3. D119249: [Attributor][NFC] Expose new API in AAPointerInfo
4. D119216: [AMDGPU] replace hostcall module flag with function attribute

Of these #1, #2 and #3 are submitted in upstream/main, while #4 is
under review.

The module flag to indicate use of hostcall is insufficient to catch
all cases where hostcall might be in use by a kernel. This is now
replaced by a function attribute that gets propagated to top-level
kernel functions via their respective call-graph.

If the attribute "amdgpu-no-hostcall-ptr" is absent on a kernel, the
default behaviour is to emit kernel metadata indicating that the
kernel uses the hostcall buffer pointer passed as an implicit
argument.

The attribute may be placed explicitly by the user, or inferred by the
AMDGPU attributor by examining the call-graph. The attribute is
inferred only if the function is not being sanitized, and the
implictarg_ptr does not result in a load of any byte in the hostcall
pointer argument.

Change-Id: I6cc12050602c3f477575c3ca09a883797169e9e3