[HeterogeneousDWARF] LLVM IR extension documentation
Document the proposed changes to LLVM IR, based on the fundamental
concepts introduced in the DWARF extension documentation.

Describe the new lifetime, fragment, and expression concepts. Document
their accompanying metadata nodes. Define the location descriptions of
several LLVM IR entities.

Demonstrate above concepts with several examples.

Change-Id: Ife58d4d43cb9cddb29153d89bb3e42abe95fa69d

[ADT] Add IntrusiveVariant, VariantTraits, and new STLForwardCompat

Add a compact, restricted version of the C++17 std::variant called
IntrusiveVariant. IntrusiveVariant stores the dynamic tag at the
beginning of every alternative type object, allowing for more efficient
packing in some cases.

Add an llvm::visit, which is compatible with std::visit and is generic
over any variant-like type which implements llvm::VariantTraits. Also
add llvm::visitSameAlternative which can be used when it is known that
all variants hold the same alternative type.

Also adds make_array, type_identity, in_place_index, and in_place_type
to STLForwardCompat.h to support the above change.

Change-Id: Ic01a1477bbec3b09085a94d8d8687ebb9be8b43a

Revert "[DebugInfo] Teach LLVM and LLDB about ptrauth in DWARF"

[HeterogeneousDWARF] Implement AMDGPU CFI, DebugInfo

Add -amdgpu-spill-cfi-saved-regs to enable complete debug info even when
codegen would otherwise not spill the EXEC mask. This is a workaround
until a solution to describe the EXEC mask without spills is integrated
based on the new structurizer.

Support emitting eh_frame for debug-info even when the target does not
support exceptions.

Extend MCAsmInfo to reflect whether a target opts in to the
heterogeneous DWARF extensions.

FIXME: Can this be split to another commit?
Teach DebugInfo about the new extension to permit testing.

Change-Id: Ic9ddc8a29d86a25a5ae5004e318cfa3e4d3ba164
Note: The encodings in BinaryFormat/Dwarf.def are subject to change.

[HeterogeneousDWARF] Add DEBUG_METADATA_VERSION_HETEROGENEOUS_DWARF

To permit development of the heterogeneous DWARF extensions in parallel
with the existing debug info, add a new debug metadata version kind.

The two schemes cannot be mixed within the same module.

Change-Id: I2a2652887bdcf67512adb7493d38cc25175c84da

[HeterogeneousDWARF][Clang] Add -gheterogeneous-dwarf

Add a frontend option to request the heterogeneous DWARF extensions.
Make this the default for AMDGCN.

Change-Id: I5be91c8dc3b8b07694d3b7ead040912e640d1e9d

[HeterogeneousDWARF] LLVM IR and Bitcode support

Implement DIExpr, DIFragment, and DILifetime metadata in IR and bitcode.
Introduce a new abstract base for DIVariable and DIFragment called
DIObject.

As a part of the introduction of DIExpr, add DIOp implemented as an
IntrusiveVariant over POD classes for each concrete operation. Separate
the implementation into an immutable DIExpr (suitable for compact
representations) and a mutable DIExprBuilder (suitable for efficiently
manipulated representations). Currently these share the same
representation (an IntrusiveVariant with a size of 64 bits).

Implement llvm.dbg.def and llvm.dbg.kill intrinsics, and DBG_DEF and
DBG_KILL pseudo machine instructions.

Add AsmParser+AsmWriter, BitcodeLoader+BitcodeWriter, and
BitcodeAnalyzer support for the above.

FIXME: Can this be split up futher? Writing meaningful tests becomes
simpler when both bitcode and IR support is present.
FIXME: Move rocm-gdb-symbols unittests into ADT unittests

Change-Id: I9c21d2ed76709ae1cbc8186281dd381d56947753

Clang codegen

Change-Id: I7da4104dab6820e5c451be4299f45c0ef5e63e5a

[HeterogeneousDWARF] Teach CloneFunction to handle new metadata

Prevent cloning of DIGlobalVariable metadata which is reachable from
function metadata when cloning a function. The identity of the global
variable is preserved in each function, unlike local variables and
fragments which are conceptually cloned along with the function.

Change-Id: I66ebb97a5b01a865432c98ee08a0114a7c689a5a

[HeterogeneousDWARF] Implement GlobalISel for llvm.dbg.{def,kill}

Translate referrers to their natural machine instruction operand, except
for allocas. Allocas are indirect (i.e. are pointers to the actual
allocation) whereas frame indices are direct (i.e. they represent the
allocation itself, not a pointer to it). To capture this in the
metadata, a level of indirection needs to be introduced in the
expression via DIExprBuilder::removeReferrerIndirection.

Change-Id: I3f4f830e968982db210198b5d17dd5777c7ddbbb

[HeterogeneousDWARF] Implement SDAGISel for llvm.dbg.{def,kill}

Much like the GlobalISel implementation, most operands are naturally
translated to machine instruction operands, whereas alloca operands
necessitate the surrounding expression be updated via
DIExprBuilder::removeReferrerIndirection.

The design of SDagISel, and the requirement that debug info does not
affect codegen, necessitates an intermediate SDDbgDefKill be maintained
until InstrEmitter. This mirrors the existing SDDbg* types used to
satisfy the same constraint for the existing debug info.

Change-Id: I2308afb0e8a2a2e56efbfedc6e5b6b2614076bee

[HeterogeneousDWARF] Implement RegAllocFast

As fast regalloc is fundamentally basic-block local, and the new debug
information at this stage is fundamentally defined across basic-blocks,
there is more involved with maintaining the new debug info than is
required for the existing debug info. Further complicating this is
that fast regalloc does not visit blocks in any particular order,
meaning the final debug info for any block cannot actually be known
until all other blocks are visited.

Implement maximal debug info for cases where a virtual register is known
spilled at the time the debug intrinsic is visited. Implement
conservatively correct debug info when the virtual register is not (yet)
spilled. In this conservative mode, record the inserted debug info and
the input debug info, and either replace it with complete info (if the
virtual register is later spilled) or retain the conservatively correct
info.

Change-Id: Iedc56604ec23dbaa447562478cdf81bb76fe1867

[HeterogeneousDWARF] Implement PrologEpilogInserter

Add a target-overridable hook to frame lowering called
insertFrameLocation, which must insert an expression which describes the
base of the current frame. This permits the generic PEI implementation
to offset the base expression to create the final expression for the
stack slot.

Change-Id: Idcd1ff66255a2c4c4f9dccc42e8483860ac25469

[HeterogeneousDWARF] Add HeterogeneousImpl for LiveDebugValues

Add a trivial implementation of LDV which transforms the
llvm.dbg.{def,kill} intrinsics into block-local constructs to simplify
the live range calculation in the AsmPrinter.

Change-Id: Id6fd30a34092d5a7f8db7bd199c3b26d568a132c

[HeterogeneousDWARF] Implement AsmPrinter at -O0

Add DbgDefKillHistoryMap to calcluate the instruction ranges over which lifetimes
are defined.

Add DwarfExprAST to lower the new DIExpr expressions.

There are several restrictions on what is currently supported, and it is
only expected that compilation at -O0 will yield correct and complete
debug information.

Change-Id: Ib80fe308ce049d71e405bbde8f4b2e27ddea6200