[HeterogeneousDWARF] Implement GlobalISel for llvm.dbg.{def,kill}
Translate referrers to their natural machine instruction operand, except
for allocas. Allocas are indirect (i.e. are pointers to the actual
allocation) whereas frame indices are direct (i.e. they represent the
allocation itself, not a pointer to it). To capture this in the
metadata, a level of indirection needs to be introduced in the
expression via DIExprBuilder::removeReferrerIndirection.

Change-Id: I3f4f830e968982db210198b5d17dd5777c7ddbbb